requirements/requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
BeautifulSoup==3.2.1
beautifulsoup4==4.5.3
bleach==2.1.3
boto==2.8.0
cairocffi==0.8.0
CairoSVG==1.0.22
certifi==2017.11.5
cffi==1.11.5
chardet==2.3.0
cssselect2==0.2.1
decorator==4.1.2
Django==1.8.13
django-bulk-update==1.1.10
django-contrib-comments==1.8.0
django-cors-headers==2.1.0
django-debug-toolbar==0.8.5
django-easy-select2==1.3.4
django-extensions==2.2.9
django-tastypie==0.12.2
djangorestframework==3.6.4
djangorestframework-csv==2.1.0
enum34==1.1.6
et-xmlfile==1.0.1
filebrowser-safe==0.4.6
future==0.16.0
gdata==2.0.15
google-api-python-client==1.6.3
grappelli-safe==0.4.5
html5lib==1.0.1
httplib2==0.10.3
idna==2.6
ijson==2.3
ipython==5.5.0
ipython-genutils==0.2.0
isodate==0.5.4
jdcal==1.3
Mezzanine==4.1.0
mysql-connector==2.2.9
#mysqlclient==1.4.6
mysql-python
newrelic==2.82.0.62
numpy==1.12.0
oauth2==1.5.211
oauth2client==4.1.2
oauthlib==2.0.1
openpyxl==2.4.7
packaging==16.8
pandas==0.19.2
pathlib2==2.3.0
pdfrw==0.4
pexpect==4.3.0
pickleshare==0.7.4
Pillow==3.2.0
poster==0.8.1
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.3.3
pyasn1-modules==0.1.1
pycparser==2.18
pyes==0.20.0
Pygments==2.2.0
pyparsing==2.2.0
Pyphen==0.9.4
python-dateutil==2.6.0
python-mimeparse==1.6.0
python-openid==2.2.5
python-social-auth==0.1.17
pytz==2016.4
requests==2.6.0
requests-oauthlib==0.3.2
rsa==3.4.2
scandir==1.6
simplegeneric==0.8.1
six==1.10.0
tinycss2==0.6.1
traitlets==4.3.2
twython==3.0.0
typing==3.7.4.1
tzlocal==1.3
unicodecsv==0.9.4
uritemplate==3.0.0
urllib3==1.6
wcwidth==0.1.7
WeasyPrint==0.42.3
webencodings==0.5
xlrd==0.9.2
XlsxWriter==0.5.1
xlwt==0.7.5
yolk==0.4.1
